move value_to_i32
