+++ crates/monty/src/types/mod.rs
@@ -63,24 +63,3 @@ pub(crate) use timedelta::TimeDelta;
 pub(crate) use timezone::TimeZone;
 pub(crate) use tuple::{Tuple, allocate_tuple};
 pub(crate) use r#type::Type;
-
-use crate::{
-    exception_private::{ExcType, RunResult, SimpleException},
-    value::Value,
-};
-
-/// Extracts an `i32` from a `Value`, accepting `Bool` and `Int`.
-///
-/// Used by `date`, `datetime`, and other constructors that expect
-/// integer arguments matching CPython's `int` coercion rules.
-pub(crate) fn value_to_i32(value: &Value) -> RunResult<i32> {
-    let int_value = match value {
-        Value::Bool(b) => i64::from(*b),
-        Value::Int(i) => *i,
-        _ => {
-            return Err(SimpleException::new_msg(ExcType::TypeError, "an integer is required (got type float)").into());
-        }
-    };
-    i32::try_from(int_value)
-        .map_err(|_| SimpleException::new_msg(ExcType::OverflowError, "signed integer is greater than maximum").into())
-}
